pyproject.toml
[tool.poetry]
name = "barky"
version = "0.1.0"
description = ""
authors = ["alvaro <garciajaenalvaro@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
notify-py = "^0.3.3"
python-telegram-bot = "^13.13"
click = "^8.1.3"

[tool.poetry.dev-dependencies]
ipython = "^8.4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
barky = 'barky.__main__:main'
